Fix items being placed into armor slots
This patch is another attempt at preventing items from being placed into a players armor slots.

I have taken a new approach and removed the previous code that was in attempt to fix this as it should no longer be needed.

By default InventoryType.PLAYER creates an inventory of size 41 (which is the players full inventory/hotbar of 36 slots, the four armor slots, and the 1 offhand slot). I have updated this to be 36 which includes only enough slots for the inventory/hotbar for a total of 36 slots.

I also updated all calls of .getContents and .setContents() to use .getStorageContents and .setStorageContents() respectively. Which returns "the contents from the section of the inventory where items can reasonably be expected to be stored." Excludes armor and result slots.

This will prevent the code from ever iterating through the four armor slots and the one offhand slot. This should fix the issue for good this time. Or at least we can hope.